:bulb: chore(#24): 리뷰반영 - 검증 항목 service 분리
